requirements.txt
telethon
openai>=1.0.0
git+https://github.com/idiap/coqui-ai-TTS.git
python-dotenv>=1.0.0